Wipmel cloud integration (#3)
* Create MELCloudACUnit_driver.groovy

Initial creation of MELCloud integration for @MatsF

* Create MELCloudParentDriver.groovy

Initial creation of MELCloud Parent drive

* Update adjustSetTemperature Logic

Change adjustSetTemperature to simply change the setTemperature attribute and not call setTemperature method as well, causing an endless loop.

* Adjust Set Temperature Change Part 2

Uncommented code in apply status updates method to call adjust set temperature and inclusion of addition debug messages in adjust set temperature method.

* Renamed method getRooms to be getStatusInfo

* Refactoring Status Info and Adjust Methods

Moving much of the conversions and other logic from apply status info method to new adjust attribute methods.

* Fixing toFloat Error

Updating temperature method to refer to parameter

* Putting value change checks into adjust methods

Including checks in adjust methods to try and avoid endless loop

* Latest updates to MELCloud AC Unit

* Fan Mode now populating

* Added initial code for set Fan Mode

* Refinements of Fan Mode Adjustments

* Changes to Temperature, Operating Mode, etc

Early copy of the code to make available for community testing

* Save AuthCode in State rather than attribute

Store authcode from Mitsubishi MelCloud in State Data for the Parent device rather than attribute.

* Thermostat and Fan Modes, Auth Code in State

Various code changes, including reading in supported thermostat and fan modes, as well as changing auth code to be stored in state data for the parent rather than an attribute on the parent.

* Fixed Typo in Min/Max Ranges for Auto

Updated the attribute names referenced in sendEvent calls when setting the min/max temperature range for auto mode.

* Fan Modes 4 and Using Numbers

Updated the Fan Modes to include Fan Mode 4 and changed the labelling of Fan Modes to be numbers instead of names, in line with screenshots from users on the HE community.

* Added Fan Mode

Based on contributions by @zimba, have added the fan() method and updated the adjustThermostatOperationMode() method to cater for fan mode and report auto mode as auto, rather than cooling.

* Fixed conversion bug

Fixed bug introduced in last commit with call to temperature conversion passing in a string when attribute value needed to be converted to a float.

* Added automatic polling to child MelCloud Driver

Added an automatic schedule for refreshing a child AC Unit, including preferences to turn the polling on and off, as well as the polling frequency.  Was trying to setup the frequency setting to be a list rather than free text, but this still needs some work.

* Fixed Error with Null Conversion

adjustSetTemperature was attempting to convert a null attribute value to a float, made us of a method to translate a null value to 23 degrees.

* Made a start of Readme.md

Started to record Readme.md for MELCloud drivers.

* Formatting changes to readme

* Additional formatting changes to Readme.md

* Formatting of Readme.md

* Update to Formatting of Readme.md

* Update README.md

* Fan Control and Fan Speed Power

Added required attributes and methods for Fan Control capability.
Added a setting to switch between text-based and number-based fan modes to support Fan tiles that have a fixed list of text-based modes for a user to select from, e.g. low, medium low, etc.
Updated setFanMode command to pass in power = true

* Align Fan Modes and Speeds, general bugs

Aligned text-based Fan Modes for Thermostat tiles to the Fan Modes displayed for a Fan tile.
Fixed a bug with setting up a new device not detecting / handling set points, etc that have not been created yet, which was resulting in toFloat being applied to a null result.