Readme.md updates, Start in folder, Auto-Inactive
Various updates to the readme for the PC Virtual Motion Sensor installation, in particular the Start-In folder setting for the Task Scheduler task and the Auto-Inactive setting for the HE device.  Many other tweaks and additions, including notes on validating the setup and adjustments to the logging once setup is complete.