Updated / Fixed Fan Speed Translation
The translation of fan modes to fan speeds that make up the Fan Control capability did not handle number-based fan modes.  This has now been updated to translate mode numbers to fan speeds.  The "Medium Low" mode in the supported mode listing had an extra space in between Medium and Low, causing the fan speed to not be set correctly.  This has now been corrected.  The Preference for number-based or text-based fan modes has been updated to more clearly explain that ON results in text-based fan modes and OFF results in number-based fan modes.