Delayed Child Device Initialization
Put a runIn call to a new initializeChildDevices method when creating devices so that the new auth code has a chance to be written after the parent refresh is complete.  The new method simply loops through child devices and calls the initialize method on those devices.  This was to get around an issue with Kumo Cloud where the same process was writing the auth code and calling initialize, but the auth code was not written until after the process completed.