Updated UT Child Driver - Temp Conversion In/Out
Added temperature conversion methods for readings coming in/out of the driver, converting those coming from the MEL/Kumo platform (in) to the chosen scale, either the HE hub's or the override value in the parent device, as well as converting those readings being sent out of the drivers to the MEL/Kumo platform.  This is partially to allow debugging of systems running Fahrenheit, but may also be useful for those wanting to operate in a different temperature scale to their hub.