Update Unified Child Driver Status Update Check
When a response for a command being issued on the Kumo platform is received, the date/time in UTC is recorded in a new lastCommandUTC attribute.  This is then compared to a date/time returned in the status response for Kumo, if the status information is older than the last command, the status update is not applied, preferring to wait until Kumo has updated.  Temperature adjustments seemed to work well, but there was still an issue when adjusting operating modes with the status reporting the old value after the mode was changed.