UT - HE Temp Scale Override (#50)
* Update Parent UT Driver - HE Scale Override 

Added heTempScale attribute and override command to allow changes in temperature scale used in drivers to override HE hub scale.

* Updated UT Child Driver - Temp Conversion In/Out

Added temperature conversion methods for readings coming in/out of the driver, converting those coming from the MEL/Kumo platform (in) to the chosen scale, either the HE hub's or the override value in the parent device, as well as converting those readings being sent out of the drivers to the MEL/Kumo platform.  This is partially to allow debugging of systems running Fahrenheit, but may also be useful for those wanting to operate in a different temperature scale to their hub.

* Update UT Readme.md - HE Temp Scale Override

Added notes in the optional configuration for the HE Temperature Scale Override command and attribute.

* UT - 1.0.7 Updated package manifest json

* Added Last Command Check for Status Updates

When a response for a command being issued on the Kumo platform is received, the date/time in UTC is recorded in a new lastCommandUTC attribute.  This is then compared to a date/time returned in the status response for Kumo, if the status information is older than the last command, the status update is not applied, preferring to wait until Kumo has updated.  Temperature adjustments seemed to work well, but there was still an issue when adjusting operating modes with the status reporting the old value after the mode was changed.

* Update Unified Child Driver Status Update Check

When a response for a command being issued on the Kumo platform is received, the date/time in UTC is recorded in a new lastCommandUTC attribute.  This is then compared to a date/time returned in the status response for Kumo, if the status information is older than the last command, the status update is not applied, preferring to wait until Kumo has updated.  Temperature adjustments seemed to work well, but there was still an issue when adjusting operating modes with the status reporting the old value after the mode was changed.

* First Commit from VS Code

* Remove old Driver files, add lint setup

* Rename MELCloud folder and cleanup MELView Drivers